requirements.txt
arrow==1.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
circuitbreaker==2.1.3
click==8.0.4
colorama==0.4.6
cryptography==45.0.7
idna==3.11
jmespath==1.0.1
oci==2.167.3
oci-cli==3.74.2
prompt-toolkit==3.0.43
pycparser==3.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.5
setuptools==80.10.2
six==1.17.0
terminaltables==3.1.10
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
